+++ CHANGELOG.md
@@ -13,6 +13,7 @@ https://github.com/nwnxee/unified/compare/build8193.16...HEAD
 - Events: added `ACTION_RESULT` to Feat/Skill/Lock events for use in the _AFTER
 - Events: added Spell Interruption events to SpellEvents
 - Events: added skippable Has Feat event to FeatEvents
+- Events: added Journal Open/Close events to JournalEvents
 - Tweaks: `NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST`
 - Tweaks: `NWNX_TWEAKS_FIX_ARMOR_DEX_BONUS_UNDER_ONE`
 - Tweaks: `NWNX_TWEAKS_FIX_ITEM_NULLPTR_IN_CITEMREPOSITORY`